+++ src/scripts/components/lists/nav-item.vue
@@ -1,22 +1,27 @@
 <template>
-  <li :class="className" :role="role" @click="handleClick">
+  <a :href="href" :class="className" @click="handleClick">
     <span class="mdc-list-item__ripple"></span>
     <slot name="before" :iconClass="UI_LIST.cssClasses.firstTile"></slot>
     <slot>
       <!-- text content -->
     </slot>
     <slot name="after" :iconClass="UI_LIST.cssClasses.lastTile"></slot>
-  </li>
+  </a>
 </template>
 
 <script>
 import UI_LIST from './constants';
 
 export default {
-  name: 'UiItem',
+  name: 'UiNavItem',
   props: {
+    // For `<router-link>`
+    href: {
+      type: String,
+      default: ''
+    },
     // States
-    selected: {
+    activated: {
       type: Boolean,
       default: false
     },
@@ -34,28 +39,9 @@ export default {
     className() {
       return {
         'mdc-list-item': true,
-        'mdc-list-item--selected': this.selected,
+        'mdc-list-item--activated': this.activated,
         'mdc-list-item--disabled': this.disabled
       };
-    },
-    role() {
-      let name = null;
-
-      if (this.$parent) {
-        switch (this.$parent.role) {
-          case 'listbox':
-            name = 'option';
-            break;
-          case 'radiogroup':
-            name = 'radio';
-            break;
-          case 'group':
-            name = 'checkbox';
-            break;
-        }
-      }
-
-      return name;
     }
   },
   methods: {